feat: moves firebase config to server/db
